requirements.txt
streamlit
snowflake-connector-python
pandas
geopandas
plotly
pillow